requirements.txt
scikit-learn
torch
prophet
pandas
numpy
joblib
scikit-learn
prophet
faiss-cpu
sentence-transformers
transformers
torch
fastapi
uvicorn
streamlit
langchain
langchain-core
langchain-community
langchain-huggingface
uvicorn